Fix [Feature Store] Schedule content doesn't match the customized data
