+++ src/components/ScheduleJob/ScheduleJob.js
@@ -27,6 +27,7 @@ import { decodeLocale, getWeekDays, getWeekStart } from '../../utils/datePicker.
 import { tabs } from './scheduleJobData'
 import { getFormatTime } from '../../utils'
 import { generateCronInitialValue } from '../../utils/generateCronInitialValue'
+import { getDefaultSchedule } from '../../utils/getDefaultSchedule'
 
 const ScheduleJob = ({
   defaultCron,
@@ -112,70 +113,7 @@ const ScheduleJob = ({
 
   useEffect(() => {
     if (defaultCron) {
-      let cron = defaultCron.split(' ')
-
-      if (cron[4] !== '*') {
-        const weekDays = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun']
-
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_ACTIVE_OPTION,
-          payload: 'week'
-        })
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_DAYS_OF_WEEK,
-          payload: cron[4].split(',').map(day => weekDays[day])
-        })
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_WEEK_TIME,
-          payload: `${cron[1] >= 10 ? cron[1] : `0${cron[1]}`}:${
-            cron[0] >= 10 ? cron[0] : `0${cron[0]}`
-          }`
-        })
-      } else if (cron[2] !== '*') {
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_ACTIVE_OPTION,
-          payload: 'month'
-        })
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_MONTH_TIME,
-          payload: `${cron[1] >= 10 ? cron[1] : `0${cron[1]}`}:${
-            cron[0] >= 10 ? cron[0] : `0${cron[0]}`
-          }`
-        })
-      } else if (cron[1] !== '*' && cron[1].match('/')) {
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_ACTIVE_OPTION,
-          payload: 'hour'
-        })
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_MINUTE,
-          payload: 0
-        })
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_HOUR,
-          payload: Number(cron[1].replace(/.*\*\//g, ''))
-        })
-      } else if (cron[1] !== '*') {
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_ACTIVE_OPTION,
-          payload: 'day'
-        })
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_DAY_TIME,
-          payload: `${cron[1] >= 10 ? cron[1] : `0${cron[1]}`}:${
-            cron[0] >= 10 ? cron[0] : `0${cron[0]}`
-          }`
-        })
-      } else {
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_ACTIVE_OPTION,
-          payload: 'minute'
-        })
-        recurringDispatch({
-          type: scheduleActionType.SCHEDULE_REPEAT_MINUTE,
-          payload: Number(cron[0].replace(/.*\*\//g, ''))
-        })
-      }
+      getDefaultSchedule(defaultCron, recurringDispatch)
     }
   }, [defaultCron])
